+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## 1.3.1 under development
 
-- no changes in this release.
+- New #657: Add `Result::add()` method for merging other results to the base one (@arogachev)
+- New #655: Add rules for validating value types - `boolean`, `float`, `integer`, `string` (@arogachev)
+- New #693: Add `AnyRule` rule (@arogachev)
+- New #649: Add `getFirstErrorMessagesIndexedByPath()` and `getFirstErrorMessagesIndexedByAttribute()` methods to
+  `Result` (@arogachev)
 
 ## 1.3.0 April 04, 2024